High Scores db final
Should be the final commit for the high scores database.
Scores are stored in ./scores.db in sqlite3. The main
important methods are startserver(), addhighscore(handle,score),
and clearhighscores(). They all do what you would expect them
to do, and if the table doesn't exist when any of them are run,
the table is created.